+++ lib/flutter_chat_ui.dart
@@ -7,20 +7,19 @@ export 'src/chat_theme.dart';
 export 'src/models/bubble_rtl_alignment.dart';
 export 'src/models/emoji_enlargement_behavior.dart';
 export 'src/models/input_clear_mode.dart';
-export 'src/models/preview_tap_options.dart';
 export 'src/models/send_button_visibility_mode.dart';
 export 'src/util.dart' show formatBytes, isConsistsOfEmojis;
-export 'src/widgets/attachment_button.dart';
 export 'src/widgets/chat.dart';
 export 'src/widgets/chat_list.dart';
-export 'src/widgets/file_message.dart';
 export 'src/widgets/image_gallery.dart';
-export 'src/widgets/image_message.dart';
-export 'src/widgets/input.dart';
-export 'src/widgets/input_text_field_controller.dart';
-export 'src/widgets/message.dart';
-export 'src/widgets/send_button.dart';
-export 'src/widgets/text_message.dart';
+export 'src/widgets/input/attachment_button.dart';
+export 'src/widgets/input/input.dart';
+export 'src/widgets/input/input_text_field_controller.dart';
+export 'src/widgets/input/send_button.dart';
+export 'src/widgets/message/file_message.dart';
+export 'src/widgets/message/image_message.dart';
+export 'src/widgets/message/message.dart';
+export 'src/widgets/message/text_message.dart';
+export 'src/widgets/message/user_avatar.dart';
+export 'src/widgets/message/user_name.dart';
 export 'src/widgets/unseen_message_banner.dart';
-export 'src/widgets/user_avatar.dart';
-export 'src/widgets/user_name.dart';
