+++ lib/src/models/pattern_style.dart
