+++ lib/src/widgets/chat.dart
@@ -13,17 +13,16 @@ import '../models/date_header.dart';
 import '../models/emoji_enlargement_behavior.dart';
 import '../models/message_spacer.dart';
 import '../models/preview_image.dart';
-import '../models/preview_tap_options.dart';
 import '../models/unseen_banner.dart';
 import '../util.dart';
 import 'chat_list.dart';
 import 'image_gallery.dart';
-import 'inherited_chat_theme.dart';
-import 'inherited_l10n.dart';
-import 'inherited_user.dart';
-import 'input.dart';
-import 'message.dart';
-import 'text_message.dart';
+import 'state/inherited_chat_theme.dart';
+import 'state/inherited_l10n.dart';
+import 'state/inherited_user.dart';
+import 'input/input.dart';
+import 'message/message.dart';
+import 'message/text_message.dart';
 import 'unseen_message_banner.dart';
 
 /// Entry widget, represents the complete chat. If you wrap it in [SafeArea] and
