+++ lib/src/widgets/input/input.dart
@@ -3,11 +3,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 
-import '../models/input_clear_mode.dart';
-import '../models/send_button_visibility_mode.dart';
+import '../../models/input_clear_mode.dart';
+import '../../models/send_button_visibility_mode.dart';
+import '../state/inherited_chat_theme.dart';
+import '../state/inherited_l10n.dart';
 import 'attachment_button.dart';
-import 'inherited_chat_theme.dart';
-import 'inherited_l10n.dart';
 import 'input_text_field_controller.dart';
 import 'send_button.dart';
