+++ lib/src/widgets/message/file_message.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 
-import '../util.dart';
-import 'inherited_chat_theme.dart';
-import 'inherited_l10n.dart';
-import 'inherited_user.dart';
+import '../../util.dart';
+import '../state/inherited_chat_theme.dart';
+import '../state/inherited_l10n.dart';
+import '../state/inherited_user.dart';
 
 /// A class that represents file message widget.
 class FileMessage extends StatelessWidget {
