+++ lib/src/widgets/message/image_message.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 
-import '../conditional/conditional.dart';
-import '../util.dart';
-import 'inherited_chat_theme.dart';
-import 'inherited_user.dart';
+import '../../conditional/conditional.dart';
+import '../../util.dart';
+import '../state/inherited_chat_theme.dart';
+import '../state/inherited_user.dart';
 
 /// A class that represents image message widget. Supports different
 /// aspect ratios, renders blurred image as a background which is visible
