+++ lib/src/widgets/message/message.dart
@@ -3,14 +3,13 @@ import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 import 'package:visibility_detector/visibility_detector.dart';
 
-import '../models/bubble_rtl_alignment.dart';
-import '../models/emoji_enlargement_behavior.dart';
-import '../models/preview_tap_options.dart';
-import '../util.dart';
+import '../../models/bubble_rtl_alignment.dart';
+import '../../models/emoji_enlargement_behavior.dart';
+import '../../util.dart';
+import '../state/inherited_chat_theme.dart';
+import '../state/inherited_user.dart';
 import 'file_message.dart';
 import 'image_message.dart';
-import 'inherited_chat_theme.dart';
-import 'inherited_user.dart';
 import 'message_status.dart';
 import 'text_message.dart';
 import 'user_avatar.dart';
