+++ lib/src/widgets/message/message_status.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 
-import 'inherited_chat_theme.dart';
+import '../state/inherited_chat_theme.dart';
 
 class MessageStatus extends StatelessWidget {
   const MessageStatus({
