+++ lib/src/widgets/message/user_avatar.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 
-import '../models/bubble_rtl_alignment.dart';
-import '../util.dart';
-import 'inherited_chat_theme.dart';
+import '../../models/bubble_rtl_alignment.dart';
+import '../../util.dart';
+import '../state/inherited_chat_theme.dart';
 
 /// Renders user's avatar or initials next to a message.
 class UserAvatar extends StatelessWidget {
