+++ lib/src/widgets/message/user_name.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 
-import '../util.dart';
-import 'inherited_chat_theme.dart';
+import '../../util.dart';
+import '../state/inherited_chat_theme.dart';
 
 /// Renders user's name as a message heading according to the theme.
 class UserName extends StatelessWidget {
