+++ lib/src/widgets/state/inherited_chat_theme.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 
-import '../chat_theme.dart';
+import '../../chat_theme.dart';
 
 /// Used to make provided [ChatTheme] class available through the whole package.
 class InheritedChatTheme extends InheritedWidget {
