+++ lib/src/widgets/state/inherited_l10n.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 
-import '../chat_l10n.dart';
+import '../../chat_l10n.dart';
 
 /// Used to make provided [ChatL10n] class available through the whole package.
 class InheritedL10n extends InheritedWidget {
